// src/config/constants/types.ts
export interface Address {
  56?: string;
  81?: string;
  82?: string;
  87?: string;
  336?: string;
  592?: string;
}
export interface Token {
  symbol?: string;
  address?: Address;
  decimals?: number;
  projectLink?: string;
  busdPrice?: string;
}
// ...
export interface FarmConfig {
  pid: number;
  lpSymbol: string;
  lpAddresses: Address;
  token: Token;
  quoteToken: Token;
  multiplier?: string;
  isCommunity?: boolean;
  dual?: {
    rewardPerBlock: number;
    earnLabel: string;
    endBlock: number;
  };
}